+++ include/pbcopper/logging/Logging.h
@@ -46,6 +46,8 @@ public:
     Logger& operator=(const Logger&) = delete;
     ~Logger();
 
+    LogLevel Level() const;
+
 private:
     // output control
     std::ofstream logFile_;
@@ -65,6 +67,8 @@ private:
     bool Handles(const LogLevel level) const;
     void MessageWriter();
 };
+
+LogLevel CurrentLogLevel();
 
 #define PBLOGGER_LEVEL(lg, lvl) \
     PacBio::Logging::LogMessage(__FILE__, __func__, __LINE__, PacBio::Logging::lvl, (lg))