+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
  - Add Sparc implementation from unanimity
+ - Add TimeReporter for conditionally logging timing info
 
 ## [1.9.1] - 2021-02-16